requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
albumentations>=1.3.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0
tqdm>=4.65.0
PyYAML>=6.0
einops>=0.7.0
segmentation-models-pytorch>=0.3.3
omegaconf>=2.3.0
requests>=2.31.0
pandas>=2.0.0
